--- .eslintcache
@@ -1,1 +1,0 @@
-[{"E:\\CodeProjects\\SUSpiciousLibraryFrontEnd\\src\\index.js":"1","E:\\CodeProjects\\SUSpiciousLibraryFrontEnd\\src\\components\\home.js":"2","E:\\CodeProjects\\SUSpiciousLibraryFrontEnd\\src\\components\\shared\\navbar.js":"3","E:\\CodeProjects\\SUSpiciousLibraryFrontEnd\\src\\api.js":"4","E:\\CodeProjects\\SUSpiciousLibraryFrontEnd\\src\\components\\book.js":"5","E:\\CodeProjects\\SUSpiciousLibraryFrontEnd\\src\\App.js":"6"},{"size":168,"mtime":1612549931127,"results":"7","hashOfConfig":"8"},{"size":234,"mtime":1612913696846,"results":"9","hashOfConfig":"8"},{"size":1629,"mtime":1612916414728,"results":"10","hashOfConfig":"8"},{"size":810,"mtime":1612549931065,"results":"11","hashOfConfig":"8"},{"size":2618,"mtime":1612908298924,"results":"12","hashOfConfig":"8"},{"size":137,"mtime":1612549931058,"results":"13","hashOfConfig":"8"},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"oev109",{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"22","messages":"23","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"E:\\CodeProjects\\SUSpiciousLibraryFrontEnd\\src\\index.js",[],"E:\\CodeProjects\\SUSpiciousLibraryFrontEnd\\src\\components\\home.js",[],"E:\\CodeProjects\\SUSpiciousLibraryFrontEnd\\src\\components\\shared\\navbar.js",[],"E:\\CodeProjects\\SUSpiciousLibraryFrontEnd\\src\\api.js",[],"E:\\CodeProjects\\SUSpiciousLibraryFrontEnd\\src\\components\\book.js",["26"],"E:\\CodeProjects\\SUSpiciousLibraryFrontEnd\\src\\App.js",[],{"ruleId":"27","severity":1,"message":"28","line":20,"column":40,"nodeType":"29","endLine":20,"endColumn":75},"jsx-a11y/alt-text","img elements must have an alt prop, either with meaningful text, or an empty string for decorative images.","JSXOpeningElement"]
