+++ src/components/Publisher/PublisherList.js
